refactor: move electron_content_utility_client to shell/app
